fix(delta): portable std::max for int64_t on Linux
next_version() did std::max(maxv /*int64_t*/, std::stoll(...) /*long long*/).
On macOS int64_t is long long so the args matched; on Linux glibc int64_t is
long, so there is no matching std::max and the build fails. Cast the stoll
result to int64_t. Surfaced when the Delta sink (in clink_sql) was first
compiled on Linux for the runtime image.

+++ include/clink/sql/delta_row_sink.hpp
@@ -183,7 +183,8 @@ private:
                         }
                     }
                     if (all_digits) {
-                        maxv = std::max(maxv, std::stoll(fn.substr(0, 20)));
+                        maxv =
+                            std::max(maxv, static_cast<std::int64_t>(std::stoll(fn.substr(0, 20))));
                     }
                 }
             }
